Un-templatize parts of the index writing code
I had recently read Modern C++ Programming when I wrote this code and
was too enthusiastic about templates. Going for the "slow" option for
now, to optimize later when it comes to that point.

This is still work-in-progress to make the indexes polymorphic:
it's hard-coding uint32 types in a few places.