Index block iterator now supports non-int types
Changed the templatized index reading code to templatize
on the type enum instead of the C++ type itself. This
makes it easier to go back and forth between the two
using type traits.

Keeping this code templatized for now since we're more
concerned about performance on the read path, and inlining
is helpful for the binary search.