Implement cfile seek-by-value
This required some fixes/changes in the StringBlockEncoder - namely,
we parse the next key ahead of what's returned. So, the first entry
which is returned by GetNextValues is the one that's already been parsed.
This is necessary during value-seeking since we need to read the value
in order to compare it against the search key.

Additionally, the private functions which deal with restart indexes
have been modified so that index 0 is the first entry in the data block,
even though it doesn't actually have an entry in the restart index array.