layer: add DeltaTrackerInterface for DeltaMemStore and DeltaFileReader
Exposing both types of delta storage under the same interface moves
towards the ability for concurrent delta flushes. The plan is to
be able to swap out a single list in order to expose a new
DeltaMemStore while the old DeltaMemStore is still active. Then,
the old DeltaMemStore can be frozen and flushed while edits are
taken into the new one. Lastly, the old DeltaMemStore is
atomic-swapped with the new DeltaFileReader.