stringbag-test: fix test failure in release mode
The test had a faulty assertion that you couldn't fit 40x6byte elements
in a 400-byte string bag. This was true in debug mode, since debug mode
has an extra debug field in the stringbag itself, but not true in
release mode.

Dropped the size of the test bag down to 300 bytes to fix the assert