Refactor row-based iterators to sit behind a common interface
Cleans up the way that memstores and layers work so that they are more
alike. This is to move towards concurrent flush/compaction so that
memstores and on-disk data can be easily swapped in place under the
same interface.

Also adds a LayerInterface in front of the Layer class, probably to be
used in the same way.