Remove use of c++11 features and some boost ptr_containers
In attempting to build with Clang 3.2, I ran into a number of issues due to
differences in implementation of the c++11 standard between gcc and clang.
GCC (at least 4.6) is based on an earlier version of the standard, and
also the version of boost on Ubuntu appears to be against that earlier version.
Building with Clang 3.2 gave a bunch of errors in the boost code itself,
in particular around ptr_containers and BOOST_FOREACH.

In the process, also fixes a number of warnings which clang emits but
gcc does not.