tablet: add the "big ugly lock" implementation of thread-safe flushes
This patch simply holds a lock during the entire flush process,
so that flushes don't overlap with other writers, etc.
This is clearly inefficient and the implementation won't last long,
just a first pass at something that works.

Also improves the test so that the flusher thread is guaranteed to run
a reasonable number of times, interleaved with the updaters.