tablet: add flag to disable duplicate key checks
This isn't a sustainable flag, but rather just a stopgap until there
are bloom filter checks against the different layers. Disabling the
duplicate key checks speeds up inserts by an order of magnitude. The
assumption is that bloom filters will have a similar effect, but
this can be used in the meantime for benchmarking.