Some simple benchmark around merging performance
Takeaway is that using a heap for merge is algorithmically
better but in practice worse than a stupid algorithm which
finds the smallest iterator at each entry through a linear
search.

May want to look into this a little further to see if it can
be optimized for the case where there are long runs where
one file is the min.