merge_iterator: fix bug when input row count is a multiple of batch size
HasNext() was returning true in the case that the previous CopyNextRows call
exactly exhausted one of the sub-iterators. The fix checks for iterator
exhaustion after each row instead of before each row.

This should fix an assertion failure seen in compaction when the input row
count was a multiple of the temporary buffer batch size.