merge_iterator: micro-optimization: avoid a vcall into source iterators
Previously saw the call into iter_->schema() show up in some profiles,
since it's called in the inner loop of the merge. This is basically
a constant so should get inlined, but was a vcall so wasn't. Now uses
the rowblock's copy of the schema, which isn't virtual.

Shaved ~2.5% off merge_iterator-test.