Implement "indirect slices" for storage of keys and values in concurrent b-tree
This should allow the b-tree to support wide rows. Somewhat work-in-progress --
still a number of optimizations that could be made. However, this increased the
average size of flushes in mt-tablet-test significantly (~10x) by making better
use of memory in the memstore. This resulted as well in a faster runtime due to
saving IO.