cbtree: fix race during insert
This bug was causing the cbtree-test to fail ~10% of the time. The
issue was that the Insert() call was changing num_entries before
adding the new entry. This caused a concurrent reader to see the
uninitialized key memory and try to interpret it as an InlineSlice,
causing a DCHECK failure in debug mode and potentially a segfault
in NDEBUG mode.