gvint: fix usages of group varint to not ever load past the end of data arrays
This was a "sort of bug" -- we were accessing uninitialized memory by using
faster group-varint code at the end of the array. The uninitialized memory was
ignored, but was still loaded into registers, so -fsanitize-address caught the
issue. The fix is to fall back to slow paths towards the end of the array.

This might slightly hurt the overall performance by adding a branch, but it
should be a very predicatable branch.