Fix cbtree to allocate nodes word-aligned
This fixes a really nasty bug whereby the atomic operations used for
concurrency control wouldn't run atomically, thereby causing failures
of cbtree-test. This was introduced when the allocation was changed to
come from arenas -- tcmalloc was allocating aligned on its own, before,
but the arena allocation isn't aligned by default.

Added aligned allocation functions to the arena, and also added assertions
to the atomic operations to check that all usages are safe.