Redo iterator interfaces to allow columnwise iteration
This is part of the work to better support late materialization of rows.
This introduces two different iterator interfaces -- row-wise (for the memstore)
and column-wise (for pretty much everything else). The supporting infrastructure
like delta tracking is now also updated to work with column-wise iterators, etc.

Both iterator types are also now batched, meaning that reading is done
by first PrepareBatch(), then Materialize{Block,Column}, then FinishBatch().
This will allow the ability to prepare a batch, only materialize some of the
columns, and then finish it without having to read the other columns off disk
in the case that none of the rows pass query predicates on the first column.