Add CollectMutations() API to delta iterator
This API will be used in the new compaction/flush code, which requires being
able to enumerate those transactions which arrive during the middle of a flush.
Additionally, it will be necessary later during a flush to be able to flush UNDO
files.

* Also adds a seeked_ member to catch misuse of the DMS iterator
  (wasted a while debugging a case where I forgot to Seek before using
  the iterator)

This code is O(n) to append mutations, but that simplifies the read path. If it becomes
a bottleneck later, we can re-evaluate.