row_changelist: add type to changelists to support DELETE and REINSERT
This support is somewhat limited at the moment -- REINSERT mutations cannot
be written to files, as a shortcut to make the implementation simpler.
This commit adds an assertion in the DeltaFile writer code to make sure
we don't accidentally run into that limitation. At some point we'll have
to deal with it, but not necessary yet.

Also adds a brief class comment for RowChangeList since it was woefully
undocumented.