Use a Row object instead of the direct uint8_t *
This patch gets rid of direct accesses to RowBlock::row_ptr(...) and
replaces a lot of "void *row" type pointers with row wrapper objects.
This is a precusor to further work which will keep the RowBlocks
in memory in a columnar layout instead of a row-based layout, which
is itself a precursor to supporting NULL bitmaps efficiently.

This has a bit of a performance regression to it, but hopefully
we'll make up the lost performance in new optimization possibilities
once the columns are contiguous in memory (eg SIMD, faster decoder
code, etc)

Signed-off-by: Todd Lipcon <todd@cloudera.com>