Support for delete in MemRowSet
Adds a Delete function to MemRowSet and allows DELETE mutations to be
stored in the mutation list. The iterator, upon seeing these mutations,
sets the corresponding row in the output RowBlock to be unselected.

If a row which has been deleted (a "ghost") is inserted again, a REINSERT
mutation is used to reset the contents of the row to the newly inserted
data.

This commit does not add support for these new mutation types in the
Flush/Compact code path.