CheckRowPresent should return false for deleted rows.
This ensures that, when a row has been deleted from a DiskRowSet, a new
insert always writes into the current MemRowSet rather than trying to
REINSERT on top of the ghost row.

This simplifies compaction a bit -- we can always garbage-collect an expired
row from a DiskRowSet without worrying that it will be REINSERTed while we're
mid-compact. (the reinsert would show up as an INSERT into the new MemRowSet
instead)