delta_tracker: keep the number of rows as a member for extra assertions
Given that a DeltaTracker is only associated with a strict non-changing
set of rows, we can pass in this count at construction time, and use
it for debug assertions that we don't try to update a row which is
outside of the bounds.

This is useful for catching bugs in compaction.