Fix bug where uncommitted mutations could get taken into account during the snapshot flush
Currently, the MemRowSetCompactionInput yields all mutations, not just those
mutations which were committed in the snapshot that was passed in.  This is to
make it more efficient, since it can just pass the MutationList from the
MemRowSet directly without any copying -- but it meant that the snapshot flush
during compaction would incorrectly include these mutations in the flushed
file.

This adds a second check during the compaction that a mutation is actually
committed before it applies it to the flushed data.

There's obviously some redundant work being done here, so left a TODO in the
code to come back and clean this up.