Change ASSERT_STATUS_OK to CHECK_OK in multithreaded tests
It turns out that gtest is not in fact fully threadsafe, and
having ASSERTs coming from multiple threads can cause SEGVs.
This changes all the assertions in multi-threaded stress
tests to CHECK macros - we don't get quite as nice error
reporting, but it should prevent flakiness in the tests,
and it should still be enough to diagnose an issue if the
test fails.