Implement an Interval Tree.
The intended use for this data structure is to store the various RowSets
that make up a tablet. With this in place, rowsets can be bounded to be
relatively small, and still have efficient access time by culling non-overlapping
files for any given query/access.

Review: http://review.sf.cloudera.com/r/29296/