Change compaction/flush to output multiple rowsets
The compaction/flush code now uses RollingDiskRowSetWriter to output
multiple RowSets, each bounded by a configurable target size. The
output files are named as before, but with a ".N" suffix indicating
their roll index. This isn't needed at load time, but is just used
to come up with unique filenames (later to be obsoleted by the
ongoing work on TabletMetadata)

Change-Id: I9d9b35408c01d4968011f0333d07d6aa9627f748
Reviewed-on: http://gerrit.ent.cloudera.com:8080/39
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins <kitchen-build@cloudera.com>