Add UBSAN build support, fix minor bugs exposed.
UBSan is a Clang option allowing to trigger crashes on undefined behavior
that might not otherwise crash. This adds a new KUDU_USE_UBSAN option
which enables it.

See http://clang.llvm.org/docs/UsersManual.html#controlling-code-generation
for more info on this sanitizer option.

This commit also makes a couple of fixes so that the UBSAN build passes:

 * Fix two cases in which &vec[0] was referenced when the vector was empty.
   In these cases, it wouldn't actually crash because the called function
   did nothing with the pointer, but it still looked dangerous and is
   apparently undefined.
 * Fix the concurrent btree which was calling methods on potentially-NULL
   objects. This worked OK because they were non-virtual methods and worked
   OK when this == NULL, but again, for sanity, we shouldn't do this.

Change-Id: I18c70c0f87af3024e6da7419877302a7d6f6261d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/43
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>