Implement some utility code necessary for budgeted compaction algorithm
This adds two algorithms important for the proposed compaction algorithm:

1) The ability to determine how "wide" a RowSet is by converting its
   string boundaries into a numeric form.

This will be used to determine the "value" of compacting away a given rowset:
if it spans a high percentage of the total tablet's key space, then it is
high value.

2) A dynamic-programming solver for the 0-1 knapsack problem.

This will be used to pick the rowsets that are most worth compacting given
a certain IO budget.

Change-Id: I1e3bc4ad0aa0e4d0b1b95d7ffbcce00ef98307a5
Reviewed-on: http://gerrit.ent.cloudera.com:8080/47
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>