Refactor CompactionPolicy to yield a set of RowSet* instead of RowSetsInCompaction
In the process of implementing the BudgetedCompaction policy, I found that
there was duplicate code around dealing with the compact_flush_lock. Also, it
was simpler to just prohibit concurrent compaction selection than to reason
about conflicting selections.

This patch addresses the above problems by moving the creation of the
RowSetsInCompaction object into Tablet::PickRowSetsToCompact and having that
method lock a tablet-scoped mutex to ensure that only one thread is making a
selection at a time.

Change-Id: I979f3487021035f55bad9cda791a5334f485076d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/82
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>