RPC race condition + minor bug fixes
This patch essentially exposes a memory allocation race condition in
rpc-test which required reworking some state handling to fix. In
rpc-test-base.h we allocate an RpcController on the stack and use it to
make synchronous calls, after which it is immediately deallocated.
Originally found this by fixing a typo in rpc-test-base.h which caused
the test timeouts to be ignored.

It turns out that there is a race where an OutboundTransfer can be in
the middle of sending data via the Reactor via Connection:WriteHandler,
write partially, then check for a timeout and execute the timeout
callback. When the send finally completes, it calls back into a
deallocated RpcController.

One solution is to move the state handling logic about the OutboundCall
from the RpcController into the OutboundCall itself. Since the
RpcController has a shared_ptr to the OutboundCall, as does the
CallTransferCallback, we avoid this issue. That's what I did here.

Also removed a couple of downcasts (double -> float, uint64 -> uint32).

Change-Id: I133b1405761e2991efbdcb25662d4023148d60f0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/51
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>