compaction: set the target rowset size in CompactionPolicy
The original compaction policy currently in master doesn't perform
very well when used with the RollingDiskRowSetWriter set to output
only small rowsets. This is because the SizeRatioBasedCompactionPolicy
tries to use rowset size as a proxy of rowset age, and when rowset rolling
is enabled, all rowsets are the same size.

This commit simply moves the setting for target rowset size into the
compaction policy itself, so that the size-ratio policy doesn't behave
poorly on the compaction branch (allowing for more fair comparisons).

Change-Id: I11e7b31d7ebc94ea808d2be32e1b8648b8e0c6be
Reviewed-on: http://gerrit.ent.cloudera.com:8080/61
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins <kitchen-build@cloudera.com>