Fix thread-safety issue with inline slices in concurrent btree
This fixes a bug that has been causing sporadic failures of mt-tablet-test
since ac49ce9e76451c1a12687d49ae13c04480f128ea, which improved string key
encoding to be a little bit denser. The bug was exposed because the
now-shorter string keys can fit in the "inline" portion of an InlineSlice,
whereas before they were always overflowing as pointers.

The cause of the bug was that we weren't treating the thread-safety of
InlineSlice carefully, and we allow concurrent writers and readers.
This allowed a reader to see an invalid pointer in an InlineSlice if
it read exactly as a writer was in the process of changing from
inlined data to indirect data.

This patch adds a new template parameter to InlineSlice to make it have
enough thread-safe semantics so as to avoid SEGVs in the concurrent
BTree use-case. There are no new unit tests, but I'm now able to loop
mt-tablet-test in an ASAN DEBUG build hundreds of times whereas it used to
fail reliably after a few iterations with an invalid memory access
or CHECK failure.

In the process, cleaned up InlineSlice in a few ways:
  * use a bitfield instead of manual bitshifting to store the
    discriminated pointer
  * removed a half-implemented feature which tried to store a prefix of
    the slice inline after the pointer

To use the atomic calls, this also required that the ConcurrentBTree
nodes align the objects on word boundaries.

Change-Id: Ia753762c477a04580cf8b2d9369568e8781013db
Reviewed-on: http://gerrit.ent.cloudera.com:8080/62
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>