Implement budget-based compaction selection
This implements the compaction selection algorithm described
in compaction-policy.txt and adds a flag to tablet.cc to
select this algorithm.

The default algorithm is not changed to 'budget' yet, pending
benchmark results. A few tests also need to be updated so that
they force full compactions on Compact() calls, since this
new algorithm will not compact non-overlapping RowSets.

Change-Id: Ie8297da6e1e4940669e2927b1674f9e1a3b551cb
Reviewed-on: http://gerrit.ent.cloudera.com:8080/84
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>