Include cyrus-sasl in build
This patch is to include Cyrus SASL (libsasl2) for Kudu's RPC system.

Kudu statically links against the libsasl2 library, however all of the
plugins for Cyrus SASL are linked dynamically. The implication is that
even though cyrus-sasl is downloaded and built as part of the thirdparty
build procedure, it (or at least the ANONYMOUS and PLAIN plugins for it)
must be available on the local system as well for the build to work and
for the tests to pass.

Some basic initialization functions are included here, as well as a very
basic unit test that ensures the library is basically working and that
the ANONYMOUS plugin is available.

Change-Id: I75645584b223d5c36874a4273a881f1476ec6601
Reviewed-on: http://gerrit.ent.cloudera.com:8080/124
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>