Add code to stream Twitter firehose via oauth, libcurl.
Since this is just a demo, and not something we'll ship, I didn't
add liboauth and libcurl as dependencies. The build will skip
building this new component if the dependencies are not available.
This means that Jenkins will not run these tests, so I ran them manually
locally. I also verified that the ingest_firehose program prints
out a bunch of JSON data from the firehose.

Change-Id: I9d9a475ba5c047eeb1df4ca175cb52b65d1b2a0a
Reviewed-on: http://gerrit.ent.cloudera.com:8080/162
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins <kitchen-build@cloudera.com>