Initial protobufs and helpers for data-path wire protocol
This patch introduces protobufs to be used on the wire for
transferring data, as well as the necessary helpers to transform
data to/from the in-memory format that the rest of Kudu
expects.

It also refactors some protobufs and helpers out of the server/
module into common/ since they will be used by both servers
and clients.

Change-Id: I475bc7ddb78334a0c1e69c62a3e07b0a93a078d5
Reviewed-on: http://gerrit.ent.cloudera.com:8080/167
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins <kitchen-build@cloudera.com>