safe_math: fix warning compiling in Clang
Clang was generating the following warning:
util/safe_math.h:20:24: warning: comparison of unsigned expression < 0 is
                        always false [-Wtautological-compare]
            ((a < 0) && (b < 0) && (a < (MathLimits<Type>::kMin - b))));

This warning was spurious, since the code appeared inside a branch which
was never taken for unsigned ints. Nevertheless, it's good to silence the
warning.

This fixes it by changing to separate specializations for signed and
unsigned.

Change-Id: I975f764e9a1978b4ac73abb172ba77df9e16cdc7
Reviewed-on: http://gerrit.ent.cloudera.com:8080/170
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>
Tested-by: jenkins <kitchen-build@cloudera.com>