tablet_server: load a demo tablet at startup
In the absence of dynamic support for creating and dropping
tablets, the tablet server now loads a tablet with the
Twitter schema at startup. This is temporary for the purposes
of demos, and will go away fairly soon.

Resubmit of change-id I99d0e8d4fea4da8a8ca945c98195fe02312e43d2 --
that one failed to build when it was submitted. This is the same
patch with only a trivial difference to build against trunk.

Change-Id: I1a764a31c4ed765af0246cf6a3ec2742bca8d79e
Originally-Reviewed-on: http://gerrit.ent.cloudera.com:8080/198
Reviewed-on: http://gerrit.ent.cloudera.com:8080/230
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>