Refactor Row classes to expose a new Cell concept, move utility code
This refactors a few functions out of RowBlockRow and ContiguousRow
to be static templated functions that can operate on any kind of row.
To do this without duplicating a bunch of other code, I added a new
Cell concept which represents a single cell within a row, exposing
methods to get/set its value and nullability, and also carrying along
its column schema information.

This was helpful in supporting AddRowToRowBlockPB() for row types
other than ConstContiguousRow -- in particular for RowBlockRows
that are the results of scans.

Change-Id: I124efa3b14104caa4da9d0180abe1fdc329373fb
Reviewed-on: http://gerrit.ent.cloudera.com:8080/211
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>