Addresses a couple of race conditions in threadpool.
There were two scenarios where we had race conditions which resulted in
deadlocks:
1- All threads are wait()'ing for work and we try to shutdown: we weren't
notifying the worker threads of the queue.clear() and both the shutdown thread
and worker threads were waiting for each other.
2 - There was an interleaved scenario work/wait for work scenario where when
we did Wait() on the ThreadPool we'd get another deadlock.

This changes thread creation not to create active threads. Threads become
active only when executing a task and become inactive right after doing so.
I ran 100K test iterations and this strategy seems to have gotten rid of the
deadlock.

Change-Id: I57b9ffad90c689cec7c0e8ae928f0ce52981920e
Reviewed-on: http://gerrit.ent.cloudera.com:8080/204
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>