memrowset: fix MaterializeBlock to initialize selection vector
This fixes an issue where the MemRowSet iterator would assume that the
selection vector had been pre-initialized to all true bits, which was
not the case. This caused incorrect results when evaluating predicates,
since any slots in the RowBlock which failed a predicate on a previous
batch would remain false for the whole iteration.

This changes the pushdown test to run three different scenarios:
all in-memory, all on-disk, and split between memory and disk.
The improved unit test fails without the fix in place.

Change-Id: Ibb847d6e9289dbb2386c50389d38d8e2b582608c
Reviewed-on: http://gerrit.ent.cloudera.com:8080/205
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>