Don't hold component_lock during compaction selection
This fixes a latency issue seen while testing the twitter demo with
a large number of rowsets in the tablet. During compaction selection,
I observed writers timing out (they are set to 1-second RPC timeouts).
Looking at the stack traces during the timeouts, I saw:

- Thread 1 (Compactor):
  - holding component_lock_ read lock
  - working on PickRowSetsToCompact (taking >1 second since there are many rowsets)
- Thread 2 (Flusher):
  - in AtomicSwapRowSets (waiting for component_lock_ write lock)
- Thread 3 (Insert RPC):
  - in Insert (waiting for component_lock_ read lock)

Even though no thread held the write lock, the Insert was blocked due to the
fairness policy of not admitting readers when there is a pending writer.

The fix is to change the compaction policy to operate on a copied reference
to the rowset tree. In order to achieve this without copying the whole
RowSetTree object, converted the member to a shared_ptr.

This also fixes an issue where the memrowset could get selected for compaction
in some unlikely corner cases. This issue was exposed by the new assertions
in this patch.

Change-Id: I36eb8557b5084f978426ff02d4e2b73db25d3f13
Reviewed-on: http://gerrit.ent.cloudera.com:8080/232
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins <kitchen-build@cloudera.com>
Reviewed-by: Alex Feinberg <afeinberg@cloudera.com>