Add TransactionContext to track inserts/mutations & add ids to MemRowSets
When recovering from failure using the WAL, we need to keep track which
records were flushed to disk (and therefore don't need to be restored)
and which ones didn't. This patch helps towards this goal by giving each
MemRowSet an unique id, which is updated in TabletSuperBlockPB each time
there is a flush. Inserts will be logged to WAL using this id (in another
patch). When reading the WAL for recovery all MemRowSets whose ids are >
than the id in TabletSuperBlockPB must be replayed.

Another requirement for recovery is that we need to keep track to which
data structures mutations are applied to. This patch makes all Mutate()
and Update() methods receive a MutationResult* that is used to
keep the ids of the MRS/Deltas. Also handles when mutations are
applied in multiple places while duplicating.

Finally, a TransactionContext keeps track of all individual Inserts()
and MutateRow() on the tablet (including the failed ones). Later on
(on another patch) TransactionContext will also keep track of the
acquired locks and keep the ScopedTransaction.

NOTE: this patch changes mem/disk row set ids to int64_t from int32_t
as we risk overflow in long running servers with high data churn.

Change-Id: Ia8ef49596deee20a5dcdd9d6a47c7e1f73046fc2
Reviewed-on: http://gerrit.ent.cloudera.com:8080/225
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>