Implement merging for minor delta compaction
This patch adapts the algorithms used for compaction of rowsets (compacion.h/.cc)
to delta files. Much of the code is copied verbatim, but per discussion
we've decided that it makes sense to have duplicate code until delta compaction
takes its final shape (later on, we'll probably want a generic merge algorithm
that can be used for both rowset and delta compactionms).

This revision does not yet implement flushing of files to disk.

Change-Id: Id1f313db1e16da6983636a598fad4e4a0b7cfed1
Reviewed-on: http://gerrit.ent.cloudera.com:8080/254
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins