Optimize the budgeted selection algorithm
This patch changes the algorithm to avoid the quadratic number of instances of
the knapsack problem. There are three key optimizations:

1) Iterate through the rowsets in a sorted order such that we don't need
   to do a second pass through the candidate list to collect rowsets
   falling within the range.

2) Because the collection of rowsets in each range now proceeds in an orderly
   fashion, we can reuse the results of each knapsack problem calculation
   and incrementally compute the new solution from the previous one with
   the new rowset added.

3) Before solving the 0-1 knapsack problem, use the solution to the fractional
   knapsack problem as an upper bound. If this upper bound is worse than the
   best solution found so far, we can skip the knapsack calculation. The
   fractional knapsack can be solved in O(n lg budget), and again can be
   done incrementally inside of existing loops.

These optimizations take the total runtime of the algorithm down from
O(n^3 * budget) to O(n^2 * budget) in the worst case, and O(n^2 log budget)
in many cases.

Because of the much improved algorithmic speed, the cost of using shared_ptrs
started to show up in the profile, so I also changed the CompactionCandidate
structure to store raw pointers instead.

To verify, I re-ran the same insert test as I used to benchmark the compaction
algorithms in the first place:

TEST_TMPDIR=/data/1/todd/kudutest GLOG_vmodule=compaction_policy=1 ./mt-tablet-test \
  --gtest_filter=MultiThreadedTabletTest/0.DoTestAllAtOnce   \
  --inserts_per_thread=125000000   \
  --num_counter_threads=0   \
  --num_insert_threads=8   \
  --num_summer_threads=0   \
  --num_slowreader_threads=0 \
  --num_updater_threads=0   \
  --flusher_backoff=1.0   \
  --flusher_initial_frequency_ms=1000   \
  --flush_threshold_mb=256   \
  --memrowset_throttle_mb=1024   \
  --tablet_compaction_policy=budget \
  --compaction_policy_dump_svgs_pattern=/tmp/budget-TIME.svg 2>&1 | tee /tmp/budget-compaction.log

Without the optimization, this inserted its last row at t=3013. With the
optimization, it inserted the last row at T=2800. This wall clock improvement may
have just been noise in the test, or may reflect the CPU savings of the new
algorithm.

The difference in write amplification was in the noise (~2% different).

The speed at which compactions were selected was much improved towards the end
of the benchmark, where the number of rowsets has increased to several hundred:

Pre-optimization:
  [todd@a1228 tmp]$ grep 'for Picking' budget-compaction.log | tail -2
  I0821 10:57:33.875324 18397 tablet.cc:406] Times for Picking rowsets to compact: real 1.860s    user 1.853s     sys 0.002s
  I0821 10:57:45.867238 18397 tablet.cc:406] Times for Picking rowsets to compact: real 1.876s    user 1.863s     sys 0.008s

With optimization:
  [todd@a1228 tmp]$ grep 'for Picking' budget-compaction.log | tail -2
  I0821 17:19:55.196177 24768 tablet.cc:406] Times for Picking rowsets to compact: real 0.029s    user 0.027s     sys 0.002s
  I0821 17:20:06.806656 24768 tablet.cc:406] Times for Picking rowsets to compact: real 0.043s    user 0.040s     sys 0.002s

Additionally ran a tablet_server process on a dataset with 700 non-overlapping
(fully compacted) rowsets. Without the change, each compaction selection took 22.2
seconds of CPU time. With the optimization, each took 0.138s of CPU time (160x speedup).

Change-Id: I82bf369a51f14aeb23da4f10531095f6b80dbb52
Reviewed-on: http://gerrit.ent.cloudera.com:8080/303
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins