Implement CFileIterator::SeekToFirst()
This revision implements a SeekToFirst() method in CFileIterator
that works irrespective of whether the cfiles use a positional or
a value index (e.g., deltas only have a value index but no
positional index).

Since this was a TODO item in cfile-dump.cc, cfile-dump.cc is also
updated to use the new method.

Change-Id: If84a03ba6a23b4a10362787e43cd366ceed30d4d
Reviewed-on: http://gerrit.ent.cloudera.com:8080/298
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>