Fix race which caused occasional segfault of mt-tablet-test
This fix is somewhat similar to 08e1788520e4586bd242b4180788ce6fa8762868:
a concurrent reader could segfault due to reading an invalid pointer
out a row in MemRowSet.

The race caused the following segfault:

*** SIGSEGV (@0x92d6e0) received by PID 10599 (TID 0x7f1f5bd88700) from PID 9623264; stack trace: ***
PC: @           0x554424 kudu::tablet::MemRowSet::Iterator::ApplyMutationsToProjectedRow()
    @           0x54fd7c kudu::tablet::MemRowSet::Iterator::MaterializeBlock()
    @           0x5b5f31 kudu::UnionIterator::MaterializeBlock()
    @           0x52f071 kudu::tablet::Tablet::Iterator::MaterializeBlock()
    @           0x50e93b kudu::RowwiseIterator::CopyBlock()
    @           0x51106f kudu::tablet::MultiThreadedTabletTest<>::CountSum()
    @           0x511463 kudu::tablet::MultiThreadedTabletTest<>::SummerThread()
    @           0x5ebc54 thread_proxy

After capturing a core dump, I could see that the ApplyMutations code was
being passed an invalid pointer for row.header_->mutation_head. But, looking
up the stack and re-evaluating the same parameter in the core file showed
a valid pointer -- this indicated that there was some race in reading this
value.

The issue turned out to be that the MRSRow value was large enough to not
fit inline in the CBTree value's InlineSlice object, and thus was allocated
out of the arena. But, it wasn't allocated aligned. Therefore, another thread
updating this same row would do a non-atomic store into the mutation_head
field, and the reader thread followed an invalid pointer.

Given that the MRSRow values are always at least 16 bytes, they're unlikely
to ever fit inline in the leaf node. So, this patch changes the leaf node
to not use the more complicated InlineSlice for values, and instead just use
a simpler class. This class always allocates the indirected data aligned
from the CBTree's arena. This then allows the use of an atomic store in
Mutation::AppendToList, which prevents the problem.

No new tests included, but without this patch, if I looped mt-tablet-test
for 12+ hours I managed to produce the segfault, and I was unable to produce
it with the patch after looping for 24h+.

Unfortunately, the change also reduces the number of entries we can fit
per leaf node, which may impact the MemRowSet performance negatively. But,
it's better to be slower and correct than fast and segfaulting. We can
revisit the performance here at a later date.

Change-Id: Ic79f9e1cc2e76290604c58f2765da984885e64e2
Reviewed-on: http://gerrit.ent.cloudera.com:8080/351
Tested-by: jenkins
Reviewed-by: Matteo Bertozzi <matteo.bertozzi@cloudera.com>