Speed up tests by allowing parallel runs
This changes jenkins to run tests in parallel, and adds properties
to our multithreaded tests which keep ctest from scheduling them
at the same time.

Also tuned down the runtime of mt-tablet-test, since it's by far the
longest-running test we have, and the DeleteAndReinsert test is
pretty stable these days.

Change-Id: Ia21d520820b396a718c64ca9d94c2bd896a078bc
Reviewed-on: http://gerrit.ent.cloudera.com:8080/348
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>