Add GetProjectionMapping() and use it in Memstore/CFileSet
Add a new Get Projection Mapping method that replaces GetProjectionFrom().
The new method will be able to handle columns not present in the base schema
and columns with different type.

MemStore, CFileSet and Deltas are already using it, but at the moment they are
not supporting the type adaptor or new columns in the projection.

Change-Id: Idcba3c7d5a9790395a1df874900fad535054c4f0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/372
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>