Enable UBSAN for jenkins ASAN builds
We should have llvm 3.3 installed on all our Jenkins slaves now, so
it's safe to enable this. I had to add a blacklist file to avoid
triggering a few false positives, and also fixed a few small places
where we were taking the address &vec[0] of an empty vector. This
wouldn't really cause issues, but triggers a UBSan error.

Change-Id: Ib0dfed17c4a558230d6557aa655133da362af804
Reviewed-on: http://gerrit.ent.cloudera.com:8080/407
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>