Add support for code coverage generation
This adds a new Jenkins build type (COVERAGE) which builds with coverage
instrumentation, runs the tests, and then runs gcovr to generate a
cobertura-compatible XML dump of the code coverage. Once this is committed
we can get this running as a regular Jenkins job.

Tested manually using:
  $ BUILD_TYPE=COVERAGE KUDU_ALLOW_SLOW_TESTS=0 ./build-support/jenkins/build-and-test.sh
on a colo box and verifying build/coverage.xml looked reasonable.

Change-Id: Ia6d8dc60b07eb08d55dd42a85a28108ea6f900aa
Reviewed-on: http://gerrit.ent.cloudera.com:8080/410
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>