Add Ids to delta file and delta mem store.
In order to perform recovery we need delta files/memstore to have
ids. The current way to id a delta file/memstore (index based on
the current number of delta), is hacky and will be broken with delta
compactions. This patch adds an id that is stored in the rowsetmetadata.
The id of the first delta memstore on Open() is the max of the current
delta files ids +1, each time a new delta memstore is created (swapped
on flush) its id is the id of the previous one +1.

Change-Id: I6adef0baff6c3280b863eb0776699a78faf660d0
Reviewed-on: http://gerrit.ent.cloudera.com:8080/423
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>