rpc: fix protobuf style from camelCase to snake_case
Per the protobuf style guide
(https://developers.google.com/protocol-buffers/docs/style) this is
the preferred way of writing fields. The Java generator is smart enough
to convert the snake_case to camelCase so if we end up using these
same proto files for a Java client we'll get the right style in both
languages.

Change-Id: I89b6516ebb369a3453ba2925dcbbcd9d5038894b
Reviewed-on: http://gerrit.ent.cloudera.com:8080/448
Reviewed-by: David Wang <dsw@cloudera.com>
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>