Add PBs for transaction results/Expose updated Tablet Metadata.
This patch replaces Operation/InsertOp/MutationOp, MutationResult etc.
with a simplified PB version.
With this at the end of a transaction the transaction context has the
full, ordered, list of all operations in PB form that can be added as
is to the commit message without having to actually go through
the updates.

In addition TabletMetadata now allows to, optionally, pass a
TabletSuperBlockPB to the update methods so that the newly created
is returned. Additionally a method to simply extract the TabletSuperBlockPB
from the current state without actually updating the state was added.

Finally this patch finalizes/integrates the format of the CommitMsg that
will be kept in the log and adds a check to TransactionContext making sure
that only inserts/mutations in the same mvcc transaction are added to a
TransactionContext.

Change-Id: I40cb37461e42471581b68f9eb46f0650291f43fa
Reviewed-on: http://gerrit.ent.cloudera.com:8080/451
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>